Updated Gravity Forms

+++ public/wp-content/plugins/gravityforms/includes/addon/class-gf-payment-addon.php
@@ -214,6 +214,8 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 			add_filter( 'gform_form_args', array( $this, 'force_ajax_for_creditcard_tokens' ), 10, 1 );
 		}
 
+		add_filter( 'gform_is_delayed_pre_process_feed', array( $this, 'maybe_delay_feed_processing' ), 20, 4 );
+
 	}
 
 	/**
@@ -393,6 +395,82 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 		parent::post_gravityforms_upgrade( $db_version, $previous_db_version, $force_upgrade );
 	}
 
+	//--------- Delayed Feeds ------
+
+	/**
+	 * Determines if feed processing is delayed by the payment feed configuration.
+	 *
+	 * @since 2.4.13
+	 *
+	 * @param bool   $is_delayed Is feed processing delayed?
+	 * @param array  $form       The form currently being processed.
+	 * @param array  $entry      The entry currently being processed.
+	 * @param string $slug       The Add-On slug e.g. gravityformsmailchimp
+	 *
+	 * @return bool
+	 */
+	public function maybe_delay_feed_processing( $is_delayed, $form, $entry, $slug ) {
+		if ( $is_delayed || ! $this->is_payment_gateway( $entry['id'] ) ) {
+			return $is_delayed;
+		}
+
+		$payment_feed = $this->current_feed;
+
+		return (bool) rgars( $payment_feed, 'meta/delay_' . $slug );
+	}
+
+	/**
+	 * Triggers processing of delayed feeds for other add-ons.
+	 *
+	 * @since 2.4.13
+	 *
+	 * @param string $transaction_id The transaction or subscription ID.
+	 * @param array  $payment_feed   The payment feed which originated the transaction.
+	 * @param array  $entry          The entry currently being processed.
+	 * @param array  $form           The form currently being processed.
+	 */
+	public function trigger_payment_delayed_feeds( $transaction_id, $payment_feed, $entry, $form ) {
+		if ( has_filter( 'gform_trigger_payment_delayed_feeds' ) ) {
+			$this->log_debug( __METHOD__ . '(): Executing functions hooked to gform_trigger_payment_delayed_feeds.' );
+
+			/**
+			 * Used in GFFeedAddOn to trigger processing of feeds delayed until payment is completed.
+			 *
+			 * @since 2.4.13
+			 *
+			 * @param string $transaction_id The transaction or subscription ID.
+			 * @param array  $payment_feed   The payment feed which originated the transaction.
+			 * @param array  $entry          The entry currently being processed.
+			 * @param array  $form           The form currently being processed.
+			 */
+			do_action( 'gform_trigger_payment_delayed_feeds', $transaction_id, $payment_feed, $entry, $form );
+		}
+	}
+
+	/**
+	 * Override to specify where the "Post Payment Action" setting should appear on the payment add-on feed.
+	 *
+	 * @since 2.4.13
+	 *
+	 * @param string $feed_slug The feed add-on slug.
+	 *
+	 * @return array
+	 */
+	public function get_post_payment_actions_config( $feed_slug ) {
+		// We specify PayPal here for backwards capability, in case the PayPal add-on < 3.3
+		// hasn't implemented get_post_payment_actions_config().
+		if ( $this->get_slug() === 'gravityformspaypal' ) {
+			$config = array(
+				'position' => 'after',
+				'setting'  => 'options',
+			);
+		} else {
+			$config = array();
+		}
+
+		return $config;
+	}
+
 	//--------- Submission Process ------
 
 	/**
@@ -471,7 +549,7 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 		// Validation called by partial entries feature via the heartbeat API.
 		$is_heartbeat = rgpost('action') == 'heartbeat';
 
-		if ( ! $validation_result['is_valid'] || ! $is_last_page || $failed_honeypot || $is_heartbeat) {
+		if ( ! $validation_result['is_valid'] || ! $is_last_page || $failed_honeypot || $is_heartbeat ) {
 			return $validation_result;
 		}
 
@@ -517,16 +595,29 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 			return $validation_result;
 		}
 
-		$submission_data = $this->get_submission_data( $feed, $form, $entry );
+		global $gf_payment_gateway;
 
-		//Do not process payment if payment amount is 0
-		if ( floatval( $submission_data['payment_amount'] ) <= 0 ) {
-
-			$this->log_debug( __METHOD__ . '(): Payment amount is zero or less. Not sending to payment gateway.' );
+		if ( $gf_payment_gateway && $gf_payment_gateway !== $this->get_slug() ) {
+			$this->log_debug( __METHOD__ . '() Aborting. Submission already processed by ' . $gf_payment_gateway );
 
 			return $validation_result;
 		}
 
+		$submission_data = $this->get_submission_data( $feed, $form, $entry );
+
+		if ( ! $this->is_valid_payment_amount( $submission_data, $feed, $form, $entry ) ) {
+			$this->log_debug( __METHOD__ . '(): Aborting. Payment amount not valid for processing.' );
+
+			return $validation_result;
+		}
+
+		if ( GFCommon::is_spam_entry( $entry, $form ) ) {
+			$this->log_debug( __METHOD__ . '() Aborting. Submission flagged as spam.' );
+
+			return $validation_result;
+		}
+
+		$gf_payment_gateway = $this->get_slug();
 
 		$this->is_payment_gateway      = true;
 		$this->current_feed            = $this->_single_submission_feed = $feed;
@@ -565,6 +656,45 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 		}
 
 		return $validation_result;
+	}
+
+	/**
+	 * Determines if the payment_amount for the current submission is valid for processing.
+	 *
+	 * @since 2.4.18
+	 *
+	 * @param array $submission_data The customer and transaction data.
+	 * @param array $feed            The feed to be processed.
+	 * @param array $form            The form being processed.
+	 * @param array $entry           The temporary entry created from the submitted values.
+	 *
+	 * @return bool
+	 */
+	public function is_valid_payment_amount( $submission_data, $feed, $form, $entry ) {
+		$is_valid = floatval( $submission_data['payment_amount'] ) > 0;
+
+		$tag      = sprintf( 'gform_%s_is_valid_payment_amount', $this->get_short_slug() );
+		$form_id  = absint( $form['id'] );
+		$tag_args = array( $tag, $form_id );
+
+		if ( gf_has_filters( $tag_args ) ) {
+			$this->log_debug( sprintf( '%s(): Executing functions hooked to %s.', __METHOD__, $tag ) );
+
+			/**
+			 * Allows custom logic to be used to determine if the add-on should process the submission for the given amount.
+			 *
+			 * @since 2.4.18
+			 *
+			 * @param bool  $is_valid        Indicates if the amount is valid for processing. Default is `true` when the amount is greater than zero.
+			 * @param array $submission_data The customer and transaction data.
+			 * @param array $feed            The feed to be processed.
+			 * @param array $form            The form being processed.
+			 * @param array $entry           The temporary entry containing the submitted values.
+			 */
+			$is_valid = (bool) gf_apply_filters( $tag_args, $is_valid, $submission_data, $feed, $form, $entry );
+		}
+
+		return $is_valid;
 	}
 
 	/**
@@ -740,6 +870,26 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 
 		return $validation_result;
 
+	}
+
+	/**
+	 * Sets the processed feed meta.
+	 *
+	 * @since 2.4.13 Overrode to prevent processed feed meta being set when a different add-on processed the submission.
+	 *
+	 * @param array $entry The Entry Object currently being processed.
+	 * @param array $form  The Form Object currently being processed.
+	 *
+	 * @return array
+	 */
+	public function maybe_process_feed( $entry, $form ) {
+		global $gf_payment_gateway;
+
+		if ( $gf_payment_gateway && $gf_payment_gateway !== $this->get_slug() ) {
+			return $entry;
+		}
+
+		return parent::maybe_process_feed( $entry, $form );
 	}
 
 	/**
@@ -1086,6 +1236,10 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 
 		$submission_data = array();
 
+		if ( empty( $feed['meta'] ) ) {
+			return $submission_data;
+		}
+
 		$submission_data['form_title'] = $form['title'];
 
 		// Getting mapped field data.
@@ -1192,7 +1346,7 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 
 		$products = GFCommon::get_product_fields( $form, $entry );
 
-		$payment_field   = $feed['meta']['transactionType'] == 'product' ? rgars( $feed, 'meta/paymentAmount' ) : rgars( $feed, 'meta/recurringAmount' );
+		$payment_field   = $this->get_payment_field( $feed );
 		$setup_fee_field = rgar( $feed['meta'], 'setupFee_enabled' ) ? $feed['meta']['setupFee_product'] : false;
 		$trial_field     = rgar( $feed['meta'], 'trial_enabled' ) ? rgars( $feed, 'meta/trial_product' ) : false;
 
@@ -1288,6 +1442,23 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 			'line_items'     => $line_items,
 			'discounts'      => $discounts
 		);
+	}
+
+	/**
+	 * Returns what should be used to prepare the payment amount; the form_total or the ID of a specific product field.
+	 *
+	 * Override if your add-on uses custom choices for the transactionType setting or does not use the standard recurringAmount and paymentAmount settings.
+	 *
+	 * @since 2.4.17
+	 *
+	 * @param array $feed The current feed.
+	 *
+	 * @return string
+	 */
+	public function get_payment_field( $feed ) {
+		$key = rgars( $feed, 'meta/transactionType' ) === 'subscription' ? 'recurringAmount' : 'paymentAmount';
+
+		return rgars( $feed, 'meta/' . $key, 'form_total' );
 	}
 
 	/**
@@ -2169,7 +2340,7 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 
 	public function get_column_value_amount( $feed ) {
 		$form     = $this->get_current_form();
-		$field_id = $feed['meta']['transactionType'] == 'subscription' ? rgars( $feed, 'meta/recurringAmount' ) : rgars( $feed, 'meta/paymentAmount' );
+		$field_id = $this->get_payment_field( $feed );
 		if ( $field_id == 'form_total' ) {
 			$label = esc_html__( 'Form Total', 'gravityforms' );
 		} else {
@@ -3223,9 +3394,11 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 		);
 		$wpdb->query( $sql );
 
-		// deleting callback log
-		$sql = $wpdb->prepare( "DELETE FROM {$wpdb->prefix}gf_addon_payment_callback WHERE addon_slug=%s", $this->_slug );
-		$wpdb->query( $sql );
+		if ( $this->_supports_callbacks ) {
+			// deleting callback log
+			$sql = $wpdb->prepare( "DELETE FROM {$wpdb->prefix}gf_addon_payment_callback WHERE addon_slug=%s", $this->_slug );
+			$wpdb->query( $sql );
+		}
 
 		//clear cron
 		wp_clear_scheduled_hook( $this->_slug . '_cron' );
@@ -3551,19 +3724,23 @@ abstract class GFPaymentAddOn extends GFFeedAddOn {
 		$form  = GFAPI::get_form( $entry['form_id'] );
 		$feed  = $this->get_payment_feed( $entry, $form );
 
+		// If user is not authorized, exit.
+		if ( $feed && $this->_slug === $feed['addon_slug'] && ! GFCommon::current_user_can_any( $this->_capabilities_settings_page ) ) {
+			wp_send_json_error( array( 'message' => esc_html__( 'Access denied.', 'gravityforms' ) ) );
+		}
+
 		//This addon does not have a payment feed. Abort.
 		if ( empty ( $feed ) ) {
 			$this->log_debug( __METHOD__ . '(): Aborting. Entry does not have a feed.' );
-
 			return;
 		}
 
 		if ( $this->cancel( $entry, $feed ) ) {
 			$this->cancel_subscription( $entry, $feed );
-			die( '1' );
+			wp_send_json_success();
 		} else {
 			$this->log_debug( __METHOD__ . '(): Aborting. Unable to cancel subscription.' );
-			die( '0' );
+			wp_send_json_error();
 		}
 
 	}
